Логика проверки авторизации  #89
- проверка данных в локальном хранилище для
автоматической авторизации
- удаление данных авторизации из локального
хранилища при нажатии Sign Out
